+++ src/modules/user/application/filters/common/NameFilter.js
@@ -1,4 +1,4 @@
-const Filter = require('../../../../../adapters/Filter');
+const Filter = require('../../../../../shared/Filter');
 
 class NameFilter extends Filter {
   constructor(value) {
